wayca scheduler daemon: make daemon accept custom config and socket path
Move the socket path definition to the lib/common.h, as it's used by
both daemon and the deployer tool, and may be used by the wayca
scheduler lib linked applications as well. So make it public.

Make the daemon accept custom config file path and socket path as well,
like the deployer tool does. This will make debug and develop easier.