lib: add prototype of wayca thread
Wayca thread is the extension of pthread. The Wayca thread
can aware the affinity information inter threads and do
the affinity binding.

Users can created wayca threads and wayca groups. Each
wayca group is composed of several wayca threads and
apply the same affinity strategy. The strategy of the
group canbe adjusted through the group attribute.

Currently the following strategy can be applied:
- topology level of the affinity
  whether the thread will be scheduled on the certain
  cpu/cluster/numa/package, etc

- thread affinity binding method
  will the thread schedule on the certain cpu or
  freely scheduled on the certain cpus.

- thread relationship in the group
  will the threads in the group be compact or scattered

The biggest advantage is that user don't need to know the
detailed information and set the cpu affinity of each
threads manually. They just need to know the relationship
and the schedule strategy of the threads, attach them
to the wayca group and set the attribute of the group.

The wayca thread is different from wayca managed thread,
that the thread created won't be managed by the wayca
scheduler daemon. The wayca managed thread should based
on wayca thread. But all the threads created by the
wayca thread, managed or not, will be know by the
wayca scheduler daemon to ensure the resouces of the
system can be finely managed by the daemon.

Still needs a lot of further implementions to enable
the whole feature, and lots of cleanups. This patch
only bring out the very basic function.

Signed-off-by: Yicong Yang <young.yicong@outlook.com>