lib: arrange wayca threads in the group according to the CPU loads
Previosly we let each group to arrange its threads according to the
CPU loads maintained by itself. For example:

Group 0's total cpuset is 0-31, and it has 8 threads with CCL scatter,
the arrangement will be like:

  0,4,8,12...

While Group 1 has the same cpuset and also 8 threads with CCL scatter,
the arrangement of group 2 will also be:

  0,4,8,12...

This will leads to the imbalance of the loads.

This patch adds a global load statistic to trace the loads of all the
cpus in this system. The 'load' here is not the so-called CPU loads
in Linux, but indicates how many wayca threads has scheduled on one
certain CPU.

For example, if there is 128 CPUs in the whole system, and one wayca
thread has scheduled on CPU 0, the loads of CPU 0 will increase 128.
Otherwise if it's not per-CPU binded and scheduled on 0-3, then the
loads will increase 128/4=32 for CPU 0-3 separately.

When the group try to figure out the CPU set for its threads,
it will consider the loads and CPU used in the group synthetically.
Take previosly case as example,

The arrangement of Group 0 will be like:

  0,4,8,12...

And for Group 1:

  1,5,9,13...

Signed-off-by: Yicong Yang <yangyicong@hisilicon.com>