lib/group: remove group attribute flag WT_GF_FIXED
The resources range of an non-empty group should be fixed
so the WT_GF_FIXED is unnecessary.

If the group is the topest level group (which has no father),
the range of its cpu resources include all the cpus in
the system. Otherwise, it's range is one topology set
according to its father's topology attribute.

Signed-off-by: Yicong Yang <yangyicong@hisilicon.com>