lib/topo: fix physical id different to logic id
physical id may different to logic id in arm. In this situation, the
physical id may be the offset of this entities in the PPTT table.

For user, this physical id is useless, and this may change in different
servers, so we should only report the logical id for user.

And more information has been printed in selftest to help developer find
more problem.

Signed-off-by: Chengchang Tang <tangchengchang@huawei.com>