lib/log: improve the log facility a bit
The WAYCA_SC_LOG_{ERR,WARN,INFO}_CONN() intends to implement
logging connected to the previous one. Currently the timestamp
is removed but the log level banner remains, get rid of it.

Noticed that wayca-sc-info used it for a non-timestamp message
logging. Provide new non-timestamp logging APIs with log level
banner and make wayca-sc-info adapted to the new APIs to avoid
broken.

Signed-off-by: Yicong Yang <yangyicong@hisilicon.com>