lib/topo: fix PCI device sysfs leaf directory
Currently, the pci device information comes from the PCI device sysfs
leaf directory. The judgment condition for whether it is a PCI device
leaf directory is whether the path contains the pci keyword. This is
very efficient, but not accurate enough.

This patch increases the judgment of the subsystem. If the subsystem
is pci, the correct leaf directory has been found.

Fixes: 9cb032286c81 ("DTS:AR0001T199 Description: lib/topo: add more topo entities and topo API")

Signed-off-by: Chengchang Tang <tangchengchang@huawei.com>