lib/topo: Fix incorrect topology information when cpu offline
Currently topology framework doesn't concern the case that CPU is offline
and may return incorrect topology information.

Fix this by checking whether the CPUs in the topology in offline or not
before returning the results. When one offline CPU is found that not
cached, then rebuild the whole topology and return the correct
information. A new function is also added to determine whether the CPU is
online or not.

Signed-off-by: Jie Liu <liujie375@h-partners.com>