tools/wayca-sc-topo: fix wayca-sc-topo memory leaks when validate format
The for loop is entered only when numa's and system's format are validated
in function add_elem_formater. In the for loop, the child of p_cont is set
to c_cont, but the parent of c_cont is not set. So unable to free memory
completely in function xmlFreeDocElementContent.
XmlDtdPtr topo_dtd is not free in function validate_format. Fix them.

Signed-off-by: Jie Liu <liujie375@h-partners.com>