tools/wayca-sc-info: fix incorrect topology printing and add physical IDs display
In past topology printing, each topology level contains a fixed number of
next topology levels by default.This is obviously wrong when the CPU is
offline.This is also not true when the CPU IDs in the core or cluster are
not contiguous.After the patch, the problem was solved.In addition, this
patch also adds the printing of physical IDs.

The following scenario is simulated in QEMU:
The entire system has two packages, each package has two nodes, each node
has two clusters, each cluster has two cores, each core has four CPUs, 64
CPUs in total.
Assume that only CPU0 of the first numa is online, all CPUs of the second
numa are offline, only CPUs of one core are online in the third numa, and
only CPUs of one cluster are online in the fourth numa.Wayca-sc-info will
print out the following topology information:

System
    Package L#0 P#36
        NUMANode P#0   mem_size 107020KB    L3_cache NA
            Cluster L#0 P#76
                Core L#0 P#236   L1i_cache NA   L1d_cache NA   L2_cache NA
                    CPU P#0
    Package L#1 P#56
        NUMANode P#2   mem_size 2023996KB   L3_cache NA
            Cluster L#1 P#156
                Core L#1 P#396   L1i_cache NA   L1d_cache NA   L2_cache NA
                    CPU P#32
                    CPU P#33
                    CPU P#34
                    CPU P#35
        NUMANode P#3   mem_size 1996092KB   L3_cache NA
            Cluster L#2 P#216
                Core L#2 P#516   L1i_cache NA   L1d_cache NA   L2_cache NA
                    CPU P#56
                    CPU P#57
                    CPU P#58
                    CPU P#59
                Core L#3 P#536   L1i_cache NA   L1d_cache NA   L2_cache NA
                    CPU P#60
                    CPU P#61
                    CPU P#62
                    CPU P#63

Signed-off-by: Jie Liu <liujie375@h-partners.com>